Conflicts:
	tests/sql/src/main/java/sql/SQLTest.java

+++ gemfirexd/shared/src/main/java/io/snappydata/thrift/common/PrepareResultHolder.java
@@ -17,7 +17,7 @@
 /*
  * Changes for SnappyData data platform.
  *
- * Portions Copyright (c) 2016 SnappyData, Inc. All rights reserved.
+ * Portions Copyright (c) 2017 SnappyData, Inc. All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you
  * may not use this file except in compliance with the License. You
@@ -52,7 +52,7 @@ public interface PrepareResultHolder {
   /**
    * the statement ID assigned by the current server
    */
-  int getStatementId();
+  long getStatementId();
 
   /**
    * update the {@link PrepareResult} e.g. after a failover
